refactor(kernfs): remove implementation of `Filesystem` to keep only utilities
